+++ database/migrations/2023_12_28_043633_create_subzones_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string("nombre");
             $table->boolean("activo");
-            $table->integer("zonas_id");
+            $table->unsignedBigInteger("zonas_id");
             $table->foreign("zonas_id")
                 ->references('id')
                 ->on("zonas")
