+++ database/migrations/2024_01_15_013725_create_posts_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string("titulo");
             $table->string("resumen");
             $table->text("contenido");
-            $table->integer("categoria_id");
+            $table->unsignedBigInteger("categoria_id");
             $table->boolean("activo");
             $table->foreign("categoria_id")
                 ->references('id')
